Workaround possible Django bug to allow admin to edit bookings.
Seems that the new 4.1 constraint validation fails on editing an object
with the type of exclusion constraint we have on bookings.

This workaround avoids python-side validation of that constraint so that
the edit can go through. If the constraint is really violated, the DB
will error on the update instead.

This is a workaround for #40 until I figure out if it's really a Django
bug or not.