Reduce access checks in booking list.
A previous commit moved the user_can_access_booking check to be
in a different place in the code where it started to be called
for every closed booking, of which there can be an awful lot. This is an
expensive check that includes DB queries each time it is run. Refactor
the code to only run it when other checks have deemed it necessary so it
runs a lot less often and hopefully results in considerably less
performance impact on users with lots of old closed bookings.